FreeRTOS-DS5
Here is an implementation of the ARM Cortex-A port on the DS-5 Cortex-A8
Fixed Virtual Platform. I was planning to do this since a long time and
now that it is finally done, you can now evaluate FreeRTOS on a Cortex-A
series processor without the need to purchase additional hardware.
Students and professionals can also explore RTOS concepts / design
methodologies or even tweak the OS, try real nasty things.

More info on:

http://rishifranklin.blogspot.in/2015/06/freertos-on-cortex-a8-fvp-arm-ds-5.html